dom/canvas/test/test_capture_throttled.html

<!DOCTYPE HTML>
<meta http-equiv="content-type" content="text/html; charset=utf-8" />

<title>Canvas2D test: CaptureStream() with throttled rAF</title>

<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="captureStream_common.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<script>
SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("Ensuring nothing happens until timing out with good margin");

const THROTTLED_THRESHOLD = 5;

// CaptureStreamTestHelper holding utility test functions.
const h = new CaptureStreamTestHelper2D();
// Canvas element captured by streams.
const c = h.createAndAppendElement('canvas', 'c');
// Video element with captureStream stream in automatic mode.
const v = h.createAndAppendElement('video', 'v');

async function checkRequestAnimationFrameThrottled() {
  const frameRate = await new Promise(resolve => {
    let start;
    let count;
    const tick = time => {
      if (!start) {
        start = time;
        count = 0;
      } else {
        count += 1;
      }
      if (time - start > 1000) {
        // One second has passed, break.
        resolve(count / ((time - start) / 1000));
        return;
      }
      window.requestAnimationFrame(tick);
    };
    window.requestAnimationFrame(tick);
  });
  ok(frameRate < THROTTLED_THRESHOLD, `rAF framerate is at ${frameRate} fps`);
}

async function checkSetTimeoutNotThrottled() {
  const start = performance.now();
  const COUNT = 5;
  for(let i = 0; i < COUNT; ++i) {
    await new Promise(resolve => setTimeout(resolve, 0));
  }
  const rate = COUNT / ((performance.now() - start) / 1000);
  ok(rate > 30, `setTimeout rate is at ${rate} ticks per second`);
}

async function checkCanvasCaptureNotThrottled() {
  const intervalMillis = 1000 / 60;
  const ctx = c.getContext('2d');
  v.srcObject = c.captureStream();
  v.play();
  const start = performance.now();
  let end;
  is(v.mozPaintedFrames, 0, "mozPaintedFrames starts at 0");
  while(true) {
    h.drawColor(c, h.green);
    await new Promise(resolve => setTimeout(resolve, intervalMillis));
    end = performance.now();
    if (end - start > 1000) {
      break;
    }
  }
  const frameRate = v.mozPaintedFrames / ((end - start) / 1000);
  ok(frameRate > THROTTLED_THRESHOLD,
    `captureStream() framerate is at ${frameRate} fps`);
}

function finish() {
  ok(true, 'Test complete.');
  SpecialPowers.wrap(window).browsingContext.isActive = true;
  SimpleTest.finish();
}

(async () => {
  // Disable background timer throttling so we can use setTimeout to draw to the
  // canvas while the refresh driver is throttled.
  await SpecialPowers.pushPrefEnv({
    set: [
      ["dom.timeout.enable_budget_timer_throttling", false],
      ["dom.min_background_timeout_value", 0],
      ["dom.min_background_timeout_value_without_budget_throttling", 0],
    ],
  });
  // Throttle the canvas' refresh driver
  SpecialPowers.wrap(window).browsingContext.isActive = false;
  await checkRequestAnimationFrameThrottled();
  await checkSetTimeoutNotThrottled();
  await checkCanvasCaptureNotThrottled();
  finish();
})();
</script>
